+++ components/BookList.tsx
@@ -1,13 +1,13 @@
 import Book, { BookProps } from '@/components/Book';
 
 interface BookListProps {
-  visible: number;
+  visibleBookCount: number;
   currPage: number;
   contents: BookProps[];
 }
 
-const BookList = ({ visible, currPage, contents }: BookListProps) => {
-  const visibleBooks = contents.slice(currPage * visible, (currPage + 1) * visible);
+const BookList = ({ visibleBookCount, currPage, contents }: BookListProps) => {
+  const visibleBooks = contents.slice(currPage * visibleBookCount, (currPage + 1) * visibleBookCount);
 
   return (
     <div className="grid grid-cols-4 gap-10 mx-8">
